fix(metro-plugin): update regex to handle non minified bundles
We did support bundles containing the following code:
```js
__d(function(g,r,i,a,m,e,d){
```

This code is produced when bundling for a release, running `./gradlew assembleRelease`
for example.

We also want to support debug variants and, in that case the code is not minified
and looks like this:

```js
__d(function (global, _$$_REQUIRE, _$$_IMPORT_DEFAULT, _$$_IMPORT_ALL, module, exports, _dependencyMap) {
```

This PR updates the regex to collect these identifiers so we can support both
variants.