feat(cli): update application sources at the same time that we create the protection
There is no need to wait for the application to be updated since we are also sending sources on `createApplicationProtection`.
Do this in parallel so can spare some time waiting for nothing.

In the future we can even move the updateApplicationSource to the backend and update them when are creating a new protection.